+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.7] - 2025-02-14
+## [1.0.8] - 2025-03-03
 
 ### Added
 
@@ -15,7 +15,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - You've got to be "interactive" with the SCCM primary site server for decryption to work 
         - This means the site server must be a client 
         - Uses script execution
-    
+   
+## [1.0.7] - 2025-02-28
+
+## Fixed
+- HTTP Module
+    - Fixed a bug where `-mp` flag wasn't correctly setting the policy request target
+
+## Added
+
+- Find module
+    - Channel binding is now supported when using NTLM auth
+- MSSQL module
+    - Channel binding is now supported when using NTLM auth
     
 ## [1.0.6] - 2024-08-15